refactor(ui): extract App.tsx module-level helpers into coddyApi, memoryTranscript, toolCallStream, reasoningTiming
Pure move of module-scope code out of App.tsx (no behavior change):
- coddyApi.ts: HDR header const, fetchJSON, randomSessionId, newId, markCoddySessionActivityRead
- chat/memoryTranscript.ts: memory copilot transcript builders + SSE phase/chunk reducers
- chat/toolCallStream.ts: tool-call SSE row types + created_at/preview readers
- chat/reasoningTiming.ts: RFC3339 parse + reasoning duration cache key
New unit tests cover the previously untested pure helpers.
