+++ external/ui/src/ui/chat/toolCallStream.ts
@@ -1,0 +1,46 @@
+export type ToolCallUpdate = {
+  toolCallId: string;
+  title?: string;
+  kind?: string;
+  status?: string;
+};
+
+export type ToolCallStatusUpdate = {
+  toolCallId: string;
+  status?: string;
+  content?: Array<{ type: string; content: { type: string; text?: string } }>;
+  _meta?: {
+    coddy?: {
+      toolResultPreview?: { truncated?: boolean; totalLines?: number };
+    };
+  };
+};
+
+export type ToolCallListRow = {
+  toolCallId: string;
+  name?: string;
+  kind?: string;
+  status?: string;
+  startedAt?: string;
+  finishedAt?: string;
+  argsPreview?: string;
+  resultPreview?: string;
+  resultPreviewTruncated?: boolean;
+  planSnapshot?: unknown;
+};
+
+export function readMessageCreatedAtUTC(
+  m: Record<string, unknown>,
+): string | undefined {
+  const raw = m.created_at ?? m.createdAt;
+  if (typeof raw !== "string") {
+    return undefined;
+  }
+  const s = raw.trim();
+  return s === "" ? undefined : s;
+}
+
+export function toolSseShowsTruncatedPreview(u: ToolCallStatusUpdate): boolean {
+  const p = u._meta?.coddy?.toolResultPreview;
+  return !!(p && p.truncated === true);
+}
